+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/KafkaBridgeConfigurationBuilder.java
@@ -18,13 +18,19 @@ import io.strimzi.api.kafka.model.common.authentication.KafkaClientAuthenticatio
 import io.strimzi.api.kafka.model.common.authentication.KafkaClientAuthenticationScramSha256;
 import io.strimzi.api.kafka.model.common.authentication.KafkaClientAuthenticationScramSha512;
 import io.strimzi.api.kafka.model.common.authentication.KafkaClientAuthenticationTls;
+import io.strimzi.api.kafka.model.common.metrics.JmxPrometheusExporterMetrics;
 import io.strimzi.api.kafka.model.common.tracing.Tracing;
+import io.strimzi.operator.cluster.model.metrics.JmxPrometheusExporterModel;
+import io.strimzi.operator.cluster.model.metrics.StrimziMetricsReporterConfig;
+import io.strimzi.operator.cluster.model.metrics.StrimziMetricsReporterModel;
 import io.strimzi.operator.common.Reconciliation;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.stream.Collectors;
 
+import static io.strimzi.api.kafka.model.common.metrics.StrimziMetricsReporter.TYPE_STRIMZI_METRICS_REPORTER;
+import static io.strimzi.operator.cluster.model.KafkaBridgeCluster.KAFKA_BRIDGE_CONFIG_VOLUME_MOUNT;
 import static io.strimzi.operator.cluster.model.KafkaBridgeCluster.OAUTH_SECRETS_BASE_VOLUME_MOUNT;
 
 /**
@@ -341,6 +347,53 @@ public class KafkaBridgeConfigurationBuilder {
             writer.println("http.producer.enabled=" + kafkaBridgeProducer.isEnabled());
         } else {
             writer.println("http.producer.enabled=true");
+        }
+
+        return this;
+    }
+
+    /**
+     * Configures the Strimzi Metrics Reporter. It is set only if user enables Strimzi Metrics Reporter.
+     *
+     * @param model     Strimzi Metrics Reporter configuration
+     *
+     * @return Returns the builder instance
+     */
+    public KafkaBridgeConfigurationBuilder withStrimziMetricsReporter(StrimziMetricsReporterModel model)   {
+        if (model != null) {
+            printSectionHeader("Strimzi Metrics Reporter configuration");
+            writer.println("bridge.metrics=" + TYPE_STRIMZI_METRICS_REPORTER);
+            // the kafka. prefix is required by the Bridge to pass Kafka client configurations
+            writer.println("kafka.metric.reporters=" + StrimziMetricsReporterConfig.KAFKA_CLASS);
+            writer.println("kafka." + StrimziMetricsReporterConfig.LISTENER_ENABLE + "=false");
+            writer.println("kafka." + StrimziMetricsReporterConfig.ALLOW_LIST + "=" + model.getAllowList());
+            writer.println();
+        }
+        return this;
+    }
+
+    /**
+     * Configures the JMX Prometheus Metrics Exporter.
+     *
+     * @param model JMX Prometheus Metrics Exporter configuration
+     * @param isLegacyMetricsConfigEnabled Flag which indicates whether the metrics are enabled or not in legacy mode.
+     *
+     * @return Returns the builder instance
+     */
+    public KafkaBridgeConfigurationBuilder withJmxPrometheusExporter(
+            JmxPrometheusExporterModel model, boolean isLegacyMetricsConfigEnabled) {
+        if (model != null || isLegacyMetricsConfigEnabled) {
+            printSectionHeader("Prometheus JMX Exporter configuration");
+            writer.println("bridge.metrics=" + JmxPrometheusExporterMetrics.TYPE_JMX_EXPORTER);
+
+            // if isLegacyMetricsConfigEnabled is not used, we pass the path of the config file.
+            // If it is used, the Bridge will use the fallback config.
+            if (!isLegacyMetricsConfigEnabled) {
+                writer.println("bridge.metrics.exporter.config.path="
+                        + KAFKA_BRIDGE_CONFIG_VOLUME_MOUNT + JmxPrometheusExporterModel.CONFIG_MAP_KEY);
+            }
+
+            writer.println();
         }
 
         return this;